Title: 사각형 채우기 3, Time: 104ms, Memory: 8MB, Status: Wrong Answer - CodeTree

+++ 240620/사각형 채우기 3/rectangle-fill-3.java
@@ -6,7 +6,7 @@ public class Main {
 
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		int n = Integer.parseInt(br.readLine());
-		int[] dp = new int[1000 + 1];
+		long[] dp = new long[1000 + 1];
 
 		dp[1] = 2;
 		dp[2] = 7;
